% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{test_nlopt_cpp}
\alias{test_nlopt_cpp}
\title{NLOPT Integration Example}
\usage{
test_nlopt_cpp()
}
\value{
A numeric value
}
\description{
A simple example for for NLOPT integration for RcppArmadillo
}
\examples{
test_nlopt_cpp()
}
